Port to Windows.
Works with catcierge_tester at least. Have not fully tested the grabber
since I lack a webcam on the windows machine I ported it on.

+++ catcierge.h
@@ -29,10 +29,10 @@ typedef struct catcierge_s
 	IplConvKernel *kernel;
 
 	int match_flipped;
-	float match_threshold;
+	double match_threshold;
 } catcierge_t;
 
-int catcierge_init(catcierge_t *ctx, const char *snout_path, int match_flipped, float match_threshold);
+int catcierge_init(catcierge_t *ctx, const char *snout_path, int match_flipped, double match_threshold);
 
 void catcierge_destroy(catcierge_t *ctx);
